+++ external/ann_1.1.1/src/kd_dump.cpp
@@ -47,7 +47,7 @@ using namespace std;					// make std:: available
 //----------------------------------------------------------------------
 
 const int		STRING_LEN		= 500;	// maximum string length
-const double	EPSILON			= 1E-5; // small number for float comparison
+const float	EPSILON			= 1E-5; // small number for float comparison
 
 enum ANNtreeType {KD_TREE, BD_TREE};	// tree types (used in loading)
